Finished DNS in Detail
Overall, DNS is a very simple concept, but it was good to know the specifics of it. I learned the process of a DNS request, hierarchies associated with domain names, types of servers that all contribute in DNS request fulfillment, and types of records that could be requested. Additionally, I did a practical simulation of DNS requests that used various record types.