Finished DFIR: An Introduction
This is a great introduction to the process of DFIR. I learned about different tools that I can use as well as two different, but similar frameworks for incident response. I learned of a few key terms pertaining to DFIR and accomplished a practical simulation to act like a digital forensics professional. I learned of the NIST SP-800-61 incident handling guide and the PICERL or the SANS incident handler's handbook